* vm_method.c (rb_gc_mark_unlinked_live_method_entries): remove unused
  variables.
	* vm_insnhelper.c (vm_call_cfunc): suppress a warning.  note that
	  `volatile type *var' doesn't make var itself volatile.

	* vm_insnhelper.c (vm_call_cfunc): remove useless hack.

	* regparse.c (onig_number_of_names): suppress a warning.

	* gc.c : remove a unused function.


git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/branches/ruby_1_9_3@39379 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ regparse.c
@@ -579,7 +579,7 @@ onig_number_of_names(regex_t* reg)
   NameTable* t = (NameTable* )reg->name_table;
 
   if (IS_NOT_NULL(t))
-    return t->num_entries;
+    return (int)t->num_entries;
   else
     return 0;
 }
